Functionality for building Clang
This entails building a modern GCC, as the GCC in Jessie is too
old to build modern Clang.

There's probably a lot of tweaking that will need to be done here.
But it is a start.

A lot of the code has been borrowed from python-build-standalone. I
was the exclusive author of said code, so there shouldn't be any
licensing concerns.